backend/requirements.txt
# Core FastAPI and dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Environment and configuration
python-dotenv>=1.0.0

# HTTP and networking
httpx>=0.25.0
requests>=2.31.0
websockets>=12.0

# External services
twilio>=8.10.0

# Rate limiting
slowapi>=0.1.9

# Database migrations
alembic>=1.13.1

# System monitoring
psutil>=5.9.6 